+++ ingest.py
@@ -1,6 +1,6 @@
 #! /usr/bin/env python3
 
-import io
+import sys
 import csv
 import json
 import requests
@@ -16,15 +16,24 @@ IDS_RULES_URL_LIST_URL = 'https://raw.githubusercontent.com/401trg/detections/ma
 
 
 def get_urls_list(url_list_url):
-    url_list_request_data = requests.get(url_list_url)
+    """ Takes in a url and returns list of urls for IOCs and IDS Rules
+    :param url_list_url: URL from 401TRG github with list of urls for IOC files and IDS files
+    :return: List of file URLS
+    """
+    url_list_request_data = requests.get(url_list_url, timeout=1)
     list_of_urls = url_list_request_data.text.splitlines()
     return list_of_urls
 
 
 def get_ioc_list(ioc_urls):
+    """ Gets a list of IOCs in Key:Value format
+    :param ioc_urls: list of urls for IOC files
+    :return: list of IOCs in Key:Value format
+    """
+
     ioc_list = list()
     for ioc_url in ioc_urls:
-        ioc_request_data = requests.get(ioc_url)
+        ioc_request_data = requests.get(ioc_url, timeout=1)
         tmp_dict = csv.DictReader(ioc_request_data.text.splitlines())
         for ioc in tmp_dict:
             ioc_list.append(json.dumps(ioc))
@@ -32,9 +41,14 @@ def get_ioc_list(ioc_urls):
 
 
 def get_ids_rules_list(ids_rules_urls):
+    """ Gets a list of IOCs in Key:Value format
+    :param ids_rules_urls: list of urls for IOC files
+    :return: list of IOCs in Key:Value format
+    """
+
     ids_rules_list = list()
     for ids_rules_url in ids_rules_urls:
-        ids_rules_request_data = requests.get(ids_rules_url)
+        ids_rules_request_data = requests.get(ids_rules_url, timeout=1)
         tmp_list = ids_rules_request_data.text.splitlines()
         ids_rules_list.extend(tmp_list)
 
@@ -42,12 +56,11 @@ def get_ids_rules_list(ids_rules_urls):
 
 
 if __name__ == "__main__":
-
     # IOC Example
     ioc_urls = get_urls_list(IOC_URL_LIST_URL)
     ioc_list = get_ioc_list(ioc_urls)
-    for ioc in ioc_list:
-        print(ioc)
+    #for ioc in ioc_list:
+    #    print(ioc)
 
     # IDS Example
     ids_rules_urls = get_urls_list(IDS_RULES_URL_LIST_URL)